Fixed few bugs in LikeController and moved dto s to other packages.
