+++ src/main/java/faang/school/postservice/dto/comment/CommentDto.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.dto.like;
+package faang.school.postservice.dto.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
