+++ src/main/java/faang/school/postservice/dto/post/PostDto.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.dto.like;
+package faang.school.postservice.dto.post;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
